refactor: move codejams endpoints to their own file
